requirements-hotstar.txt
# Python dependencies for Hotstar ingestion scripts

# Database
psycopg2-binary==2.9.9

# HTTP requests
requests==2.31.0

# Cloud SQL connector (for Google Cloud)
cloud-sql-python-connector[pg8000]==1.6.0
pg8000==1.30.4

# Google Cloud Secret Manager
google-cloud-secret-manager==2.18.0
